refactor: move contributor exemptions to contributor routes
