+++ package.json
@@ -27,7 +27,7 @@
     "axios": "^1.6.7",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
-    "express": "^4.18.2",
+    "express": "^5.0.0-beta.3",
     "express-validator": "^7.0.1",
     "helmet": "^7.1.0",
     "http-status": "^1.7.4",
