+++ src/services/CompanyService.ts
@@ -3,9 +3,7 @@ import { prisma } from "@/prisma";
 export async function getCompanyByUserId(userId: number) {
   return await prisma.company.findFirst({
     where: {
-      userId: {
-        equals: userId,
-      },
+      userId,
     },
   });
 }
